+++ relay/polkadot/src/lib.rs
@@ -215,7 +215,6 @@ impl Contains<RuntimeCall> for PostAhmFilter {
 			FastUnstake(..) |
 			Slots(..) |
 			Auctions(..) |
-			StateTrieMigration(..) |
 			AssetRate(..) => false,
 
 			// Session keys are managed via Asset Hub post-AHM (forwarded to the relay through
@@ -1785,27 +1784,6 @@ impl ah_client::SendToAssetHub for StakingXcmToAssetHub {
 	}
 }
 
-parameter_types! {
-	// The deposit configuration for the singed migration. Specially if you want to allow any signed account to do the migration (see `SignedFilter`, these deposits should be high)
-	pub const MigrationSignedDepositPerItem: Balance = CENTS;
-	pub const MigrationSignedDepositBase: Balance = 20 * CENTS * 100;
-	pub const MigrationMaxKeyLen: u32 = 512;
-}
-
-impl pallet_state_trie_migration::Config for Runtime {
-	type RuntimeHoldReason = RuntimeHoldReason;
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type SignedDepositPerItem = MigrationSignedDepositPerItem;
-	type SignedDepositBase = MigrationSignedDepositBase;
-	type ControlOrigin = EnsureRoot<AccountId>;
-	type SignedFilter = frame_support::traits::NeverEnsureOrigin<AccountId>;
-
-	// Use same weights as substrate ones.
-	type WeightInfo = pallet_state_trie_migration::weights::SubstrateWeight<Runtime>;
-	type MaxKeyLen = MigrationMaxKeyLen;
-}
-
 /// The [frame_support::traits::tokens::ConversionFromAssetBalance] implementation provided by the
 /// `AssetRate` pallet instance.
 ///
@@ -1934,9 +1912,6 @@ construct_runtime! {
 		Crowdloan: crowdloan = 73,
 		Coretime: coretime = 74,
 
-		// State trie migration pallet, only temporary.
-		StateTrieMigration: pallet_state_trie_migration = 98,
-
 		// Pallet for sending XCM.
 		XcmPallet: pallet_xcm = 99,
 
@@ -1999,12 +1974,25 @@ pub type TxExtension = (
 pub mod migrations {
 	use super::*;
 
+	frame_support::parameter_types! {
+		pub const StateTrieMigrationName: &'static str = "StateTrieMigration";
+	}
+
+	/// Remove the `StateTrieMigration` pallet's storage. The state trie migration on Polkadot is
+	/// complete and the pallet has been removed from the runtime, see
+	/// <https://github.com/polkadot-fellows/runtimes/issues/905>.
+	pub type RemoveStateTrieMigrationPallet = frame_support::migrations::RemovePallet<
+		StateTrieMigrationName,
+		<Runtime as frame_system::Config>::DbWeight,
+	>;
+
 	/// Unreleased migrations. Add new ones here:
 	pub type Unreleased = (
 		parachains_on_demand::migration::MigrateV1ToV2<Runtime>,
 		parachains_scheduler::migration::MigrateV3ToV4<Runtime>,
 		parachains_configuration::migration::v13::MigrateToV13<Runtime>,
 		parachains_shared::migration::MigrateToV2<Runtime>,
+		RemoveStateTrieMigrationPallet,
 	);
 
 	/// Migrations/checks that do not need to be versioned and can run on every update.